refactored grunt tasks into folders
