+++ grunt/markdox.js
@@ -1,0 +1,35 @@
+'use strict';
+
+module.exports = function (grunt) {
+    grunt.loadNpmTasks('grunt-markdox');
+    grunt.config.set('markdox', {
+        options: {
+            // Task-specific options go here.
+            template: 'documentation/template.ejs'
+        },
+        target: {
+            files: [
+                // {
+                //     src: 'src/service/configuration-service.js', dest: 'documentation/generated/configuration-service.html.md'
+                // },
+                {
+                    src: 'src/service/run-api-service.js',
+                    dest: 'documentation/generated/run-api-service/index.html.md'
+                }, {
+                    src: 'src/service/data-api-service.js',
+                    dest: 'documentation/generated/data-api-service/index.html.md'
+                }, {
+                    src: 'src/service/auth-api-service.js',
+                    dest: 'documentation/generated/auth-api-service/index.html.md'
+                }, {
+                    src: 'src/service/variables-api-service.js',
+                    dest: 'documentation/generated/variables-api-service/index.html.md'
+                }, {
+                    src: 'src/managers/run-manager.js',
+                    dest: 'documentation/generated/run-manager/index.html.md'
+                }
+
+            ]
+        }
+    });
+};
