+++ grunt/watch.js
@@ -1,0 +1,15 @@
+'use strict';
+
+module.exports = function (grunt) {
+    grunt.loadNpmTasks('grunt-contrib-watch');
+    grunt.config.set('watch', {
+        source: {
+            files: ['src/**/*.js'],
+            tasks: ['browserify2:edge', 'mocha:test']
+        },
+        tests: {
+            files: ['tests/spec/**/*.js'],
+            tasks: ['mocha:test']
+        }
+    });
+};
