+++ package.json
@@ -8,6 +8,7 @@
   "devDependencies": {
     "growl": "^1.8.1",
     "grunt": "^0.4.5",
+    "grunt-browserify": "^3.3.0",
     "grunt-browserify2": "^0.1.8",
     "grunt-bump": "0.0.15",
     "grunt-contrib-jshint": "^0.10.0",
@@ -17,12 +18,11 @@
     "grunt-jscs": "^0.8.1",
     "grunt-jscs-checker": "^0.6.1",
     "grunt-markdox": "^1.2.1",
-    "grunt-mocha-phantom-istanbul": "^0.1.0",
+    "grunt-mocha-phantom-istanbul": "^0.1.1",
     "husky": "^0.5.1",
-    "jshint-stylish": "^0.2.0",
-    "uglify-js": "^2.4.15",
     "istanbul": "^0.3.5",
-    "grunt-mocha-phantom-istanbul": "^0.1.1"
+    "jshint-stylish": "^0.2.0",
+    "uglify-js": "^2.4.15"
   },
   "scripts": {
     "prepush": "grunt validate"
